GG-48489 .NET: Enable multi-targeting (#3664)
- Multi-target the libraries `Apache.Ignite.Core` and `Apache.Ignite.Linq` to `netstandard2.0;net8.0;net10.0`
- Fix `build.ps1` to publish each production project once per declared TFM into `bin\<tfm>`; replace solution-based build
- Fix the code to compile on new targets (suppress warnings, gate legacy logic)

Multi-targeting allows us to use modern language features while keeping backwards compatibility for existing users on older targets.

+++ modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryReflectiveActions.cs
@@ -261,11 +261,11 @@ namespace Apache.Ignite.Core.Impl.Binary
             else if (type == typeof(UIntPtr))
             {
                 writeAction = raw
-                    ? GetRawWriter<UIntPtr>(field, (w, o) => w.WriteLong((long) o))
-                    : GetWriter<UIntPtr>(field, (f, w, o) => w.WriteLong(f, (long) o));
+                    ? GetRawWriter<UIntPtr>(field, (w, o) => w.WriteLong((long) (ulong) o))
+                    : GetWriter<UIntPtr>(field, (f, w, o) => w.WriteLong(f, (long) (ulong) o));
                 readAction = raw
-                    ? GetRawReader(field, r => (UIntPtr) r.ReadLong())
-                    : GetReader(field, (f, r) => (UIntPtr) r.ReadLong(f));
+                    ? GetRawReader(field, r => (UIntPtr) (ulong) r.ReadLong())
+                    : GetReader(field, (f, r) => (UIntPtr) (ulong) r.ReadLong(f));
             }
             else
             {
